+++ CHANGELOG.md
@@ -3,9 +3,19 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic 
+and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- support for WalkingPad A1 Pro
+- manual config flow
+- bluetooth discovery config flow
+- distance sensor
+- steps sensor
+- duration sensors (available in several units : minutes, hours, days)
+- current speed sensor
 
 [unreleased]: https://github.com/madmatah/hass-walkingpad